Reverse alternate nodes and append to the end
Given a linked list, reverse alternate nodes and append them to the end of the list. Extra allowed space is O(1)

Examples:

Input: 1->2->3->4->5->6
Output: 1->3->5->6->4->2
Explanation: Two lists are 1->3->5 and 2->4->6, 
reverse the 2nd list: 6->4->2. 
Merge the lists 

Input: 12->14->16->18->20
Output: 12->16->20->18->14
Explanation: Two lists are 12->16->20 and 14->18, 
reverse the 2nd list: 18->14. 
Merge the lists